By Bhargav Kumar
Objective: The objective of this project is to understand the UK bank customer data.


The description of the dataset is mentioned below:
Data source: https://www.kaggle.com/ukveteran/uk-bank-customers
Use the default data set (.csv file) available. 

Problem statements:
Based on the .csv file provided, we are going to analyze and develop a dashboard to 
understand the customer data of a UK bank. Use filters, parameters and actions 
wherever possible to make the dashboard interactive.
 Create an asymmetric drop down of Region with their respective customer names 
and their Balances with a gender wise color code. Use set actions to complete this 
task.
 A region wise bar chart which displays the count of customers based on High and 
low balance. Create a parameter to let the users’ dynamically decide the limit 
value of balance which categorizes it into high and low. 
 Create a donut pie chart which displays a percent of total balances in a region 
wise manner.
 Display a forecast with a day wise balance chart using Exponential smoothing.
 Create a stacked bar chart to view the gender wise average age in every region.
 Analyze the distribution of a measure (Balance) using a Histogram and create a 
parameter to let the users’ dynamically decide the bin size.
 Compare the month wise balance using a motion chart with marks and trails.
 Create a moving average balance line chart, average calculated for a period of past 
2 weeks. Compare it with the actual values.
 Compare the region wise average balance with the (region wise average balance) 
in-terms of its job classifications.
 Region wise pie chart displaying the job classifications with respect to its 
balances.
 Create a global date range parameter.
 Include interactive filters for Job classifications and Highlighters for Region in the 
final dashboard.