Add files via upload
Objective: The objective of this project is to understand the global covid-19 mortality 
rates. 

The description of the dataset is mentioned below. 
Use the default data-sets-1, data-set-2 (.csv files) available. Data sets are to be related, joined or 
blended to solve the following problem statements.

Problem statement:
Based on the csv. files for global mortality rates, we are going to analyze and develop a 
dashboard to understand the covid-19 global cases. Use filters, parameters and actions 
wherever possible to make the dashboard interactive.
 Comparing the global confirmed vs. death cases in a world map using pie charts.
 Top 5 countries based on confirmed and death cases in numbers.
 Create a parameter for percentile comparison between countries based on 
confirmed or death cases.
 Comparing the country wise cases using logarithmic axes. Dashboard should 
display both log axis chart and a default axis chart in an alternate interactive way.
 New cases per day in China and India – compared in a date wise chart.
 Which day has the highest new death cases?
 Which WHO region has the highest new cases verses new deaths ratio?
 Create a parameter to dynamically view Top N WHO regions based on cumulative 
new cases and death cases ratio.
 Average WHO region wise cumulative cases to be visualized using a funnel chart.
 Dashboard should have a drop-down menu to view the WHO region wise data 
using a bar chart, line chart or a map as per user’s requirement.