By Bhargav Kumar
Objective: The objective of this project is to understand the agricultural data.

Use the default Agriculture-data set (excel file) available. 

Problem statements:
Based on the excel file provided, we are going to analyze and develop a dashboard to 
understand the area, yield and production from the data. Use filters, parameters and 
actions wherever possible to make the dashboard interactive.
 An interactive tree map to display district wise data. Tree map should have state 
labels. On hovering on a particular state, the corresponding districts data are to be 
displayed.
 An asymmetric drill down to display the list of crops under each category.
 Add URL actions, which direct users’ to a Google search page of the selected crop. 
Web page is to be displayed the final dashboard.
 A packed bubble chart to display the list of crops cultivated throughout the year.
 Under the season Kharif, display the top 2 crop categories. Hint: Use context 
filters
 Create a hierarchy of seasons, crop categories and the list of crops under each. Add 
highlighters for season.
 One major sheet in the final dashboard should be unaffected by any action applied. 
Use the view in this major sheet to filter data in the other.
 Using parameters color code, the seasons with high yield and low yield based on 
its crop categories.
 Season wise crop category production pie charts for districts cultivating 
throughout the year.
 Rank the crops based on their yield.
 Create a chart to display crops with respect to their yield measuring units.